requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.2.0
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pydantic==1.10.7 